+++ ccramic/io/pdf.py
@@ -3,8 +3,8 @@ from matplotlib.backends.backend_pdf import PdfPages
 
 from ccramic.utils.cell_level_utils import (
     get_min_max_values_from_zoom_box,
-    get_min_max_values_from_rect_box,
-    get_bounding_box_for_svgpath)
+    get_min_max_values_from_rect_box)
+from ccramic.utils.pixel_level_utils import get_bounding_box_for_svgpath
 import numpy as np
 import cv2
 from PIL import Image